[refactor] Long으로 변경.

+++ src/main/java/community/whatever/onembackendjava/UrlMapping.java
@@ -4,11 +4,11 @@ import java.time.Instant;
 
 record UrlMapping(String originalUrl, Instant expiryTime) {
 
-    public UrlMapping(String originalUrl, Integer ttlMinutes) {
+    public UrlMapping(String originalUrl, Long ttlMinutes) {
         this(originalUrl, calculateExpiryTime(ttlMinutes));
     }
     
-    private static Instant calculateExpiryTime(Integer ttlMinutes) {
+    private static Instant calculateExpiryTime(Long ttlMinutes) {
         long ttlSeconds = ttlMinutes * 60;
         return Instant.now().plusSeconds(ttlSeconds);
     }
